+++ src/template/js/src/client/util.js
@@ -1,2 +1,0 @@
-export const isEqual = (...objects) =>
-  objects.every((obj) => JSON.stringify(obj) === JSON.stringify(objects[0]));