+++ src/util.ts
@@ -5,33 +5,41 @@ import path from "path";
 import semver from "semver";
 import validateProjectName from "validate-npm-package-name";
 
+import { FileCopy } from "./types";
+
 /**
  * Executes a command in a spawned process.
  * @param command Command to execute in the process.
  * @param args Additional arguments to attach to the command.
  * @param options Optional options object to pass along.
+ * @param debug Optional logging of output.
  */
 export const executeCommand = async (
   command: string,
   args?: string[],
-  options?: { cwd?: string; shell?: boolean }
+  options?: { cwd?: string; path?: string; shell?: boolean },
+  debug?: boolean
 ): Promise<void | { code: number; signal: any }> =>
   new Promise((resolve, reject) => {
     const cp = spawn(command, args, options);
+    if (debug) {
+      cp.stdout.on("data", (data) => {
+        console.log(`stdout: ${data}`);
+      });
+
+      cp.stderr.on("data", (data) => {
+        console.error(`stderr: ${data}`);
+      });
+    }
+
     cp.on("error", (err: Error) => {
       if (err) {
-        console.log("");
-        console.error(JSON.stringify({ err, command, args, options }, null, 2));
         console.log("");
         reject(err.message);
       }
     });
     cp.on("exit", (code: number | null, signal) => {
       if (code !== 0) {
-        console.log("");
-        console.error(
-          JSON.stringify({ args, command, code, signal, options }, null, 2)
-        );
         console.log("");
         reject({ args, command, code, signal, options });
       }
@@ -75,6 +83,7 @@ export const cleanupError = async (
  * @param replaceToken Regex used to find the values in the template file.
  * @param applicationName Name of application.
  * @param authorName Name of author.
+ * TODO - Create a better version of this function - very similar to "replacer", but smaller.
  */
 export const valueReplacer = (
   files: string[],
@@ -133,4 +142,60 @@ export const verifyNodeVersion = (): void => {
     console.error(chalk.red(`Please update your version of Node.`));
     process.exit(1);
   }
+};
+
+/**
+ * Generates an array of files to copy during the template process.
+ * @param dir __dirname.
+ * @param root Root application directory.
+ * @param language Language of application.
+ */
+export const generateFilesToCopyArr = (
+  dir: string,
+  root: string,
+  language: "js" | "ts"
+): FileCopy[] => {
+  const filesToCopy = [
+    {
+      src: path.join(dir, `template/src`),
+      dest: path.join(root, "/src"),
+    },
+    {
+      src: path.join(dir, `template/public`),
+      dest: path.join(root, "/public"),
+    },
+    {
+      src: path.join(dir, `template/env-cmdrc.json`),
+      dest: path.join(root, "/.env-cmdrc.json"),
+    },
+    {
+      src: path.join(dir, "template/README.md"),
+      dest: path.join(root, "/README.md"),
+    },
+    {
+      src: path.join(dir, "template/gitignore"),
+      dest: path.join(root, "/.gitignore"),
+    },
+    {
+      src: path.join(dir, `template/webpack-${language}.js`),
+      dest: path.join(root, "/webpack.config.js"),
+    },
+    {
+      src: path.join(dir, `template/template-tsconfig.json`),
+      dest: path.join(root, "/template-tsconfig.json"),
+    },
+    {
+      src: path.join(dir, `template/index.d.ts`),
+      dest: path.join(root, "/index.d.ts"),
+    },
+  ];
+
+  // * Copy .babelrc for JS projects
+  if (language === "js")
+    filesToCopy.push({
+      src: path.join(dir, "template/babelrc"),
+      dest: path.join(root, "/.babelrc"),
+    });
+
+  return filesToCopy;
 };